Continue type alias purge - fixing more files
- Fixed tty.c remaining dword_t uses
- Fixed task.c dword_t and int_t uses
- Fixed stat.c function signatures
- Fixed calls.h stat function declarations
- Fixed inode.h flock structs
- Fixed real.h, real.c
- Fixed sock.c socket functions
- Fixed signal.c, aarch64/signal.c
- Fixed ptrace.c
- Fixed random.c
- Updated signal.h declarations

+++ Sources/IXLandLinuxRuntime/fs/sock.h
@@ -12,8 +12,8 @@
 int64_t sys_socketcall(uint32_t call_num, addr_t args_addr);
 
 int64_t sys_socket(uint32_t domain, uint32_t type, uint32_t protocol);
-int64_t sys_bind(fd_t sock_fd, addr_t sockaddr_addr, uint64_t sockaddr_len);
-int64_t sys_connect(fd_t sock_fd, addr_t sockaddr_addr, uint64_t sockaddr_len);
+int64_t sys_bind(fd_t sock_fd, addr_t sockaddr_addr, uint32_t sockaddr_len);
+int64_t sys_connect(fd_t sock_fd, addr_t sockaddr_addr, uint32_t sockaddr_len);
 int64_t sys_listen(fd_t sock_fd, int64_t backlog);
 int64_t sys_accept(fd_t sock_fd, addr_t sockaddr_addr, addr_t sockaddr_len_addr);
 int64_t sys_getsockname(fd_t sock_fd, addr_t sockaddr_addr, addr_t sockaddr_len_addr);
